feat: Easier way to create a Formula & Better conception for Model estimation
