+++ powerapi/handler/__init__.py
@@ -32,3 +32,4 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from powerapi.handler.handler import Handler, InitHandler, HandlerException
 from powerapi.handler.poison_pill_message_handler import PoisonPillMessageHandler
 from powerapi.handler.start_handler import StartHandler
+from powerapi.handler.formula_handler import FormulaHandler
